Rename m0_state.{c,h} to usb_api_m0_state.{c,h}
