+++ firmware/hackrf_usb/usb_api_m0_state.c
@@ -19,7 +19,7 @@
  * Boston, MA 02110-1301, USA.
  */
 
-#include "m0_state.h"
+#include "usb_api_m0_state.h"
 
 #include <libopencm3/lpc43xx/sgpio.h>
 #include <stddef.h>
