+++ firmware/hackrf_usb/usb_api_m0_state.h
